+++ packtools/domain.py
@@ -116,6 +116,70 @@ def XSLT(xslt_name):
         return xslt
 
 
+#----------------------------------
+# validators for etree._ElementTree
+#
+# a validator is an object that
+# provides the method ``validate``,
+# with the following signature:
+# validate(xmlfile: etree._ElementTree) -> Tuple(bool, list)
+#---------------------------------- 
+class PyValidator(object):
+    """Style validations implemented in Python.
+    """
+    def __init__(self, pipeline=checks.StyleCheckingPipeline):
+        self.ppl = pipeline()
+
+    def validate(self, xmlfile):
+        errors = next(self.ppl.run(xmlfile, rewrap=True))
+        return bool(errors), errors
+
+
+class DTDValidator(object):
+    """DTD validations.
+    """
+    def __init__(self, dtd):
+        self.dtd = dtd
+
+    def validate(self, xmlfile):
+        """Validate xmlfile against the given DTD.
+
+        Returns a tuple comprising the validation status and the errors list.
+        """
+        result = self.dtd.validate(xmlfile)
+        errors = [style_errors.SchemaStyleError(err)
+                  for err in self.dtd.error_log]
+
+        return result, errors
+
+
+class SchematronValidator(object):
+    """Style validations implemented in Schematron.
+    """
+    def __init__(self, sch):
+        self.sch = sch
+
+    @classmethod
+    def from_catalog(cls, ref):
+        """Get an instance based on schema's reference name.
+
+        :param ref: The reference name for the schematron file in 
+                    :data:`packtools.catalogs.SCH_SCHEMAS`.
+        """
+        return cls(StdSchematron(ref))
+
+    def validate(self, xmlfile):
+        """Validate xmlfile against the given Schematron schema.
+
+        Returns a tuple comprising the validation status and the errors list.
+        """
+        result = self.sch.validate(xmlfile)
+        errors = [style_errors.SchematronStyleError(err)
+                  for err in self.sch.error_log]
+
+        return result, errors
+
+
 #--------------------------------
 # adapters for etree._ElementTree
 #--------------------------------
@@ -131,21 +195,25 @@ class XMLValidator(object):
     :param file: etree._ElementTree instance.
     :param sps_version: the version of the SPS that will be the basis for validation.
     :param dtd: (optional) etree.DTD instance. If not provided, we try the external DTD.
-    :param sch_schemas: (optional) list of Schematron schemas.
+    :param style_validators: (optional) list of
+                             :class:`packtools.domain.SchematronValidator`
+                             objects.
     """
-    def __init__(self, file, dtd=None, sch_schemas=None):
+    def __init__(self, file, dtd=None, style_validators=None):
         assert isinstance(file, etree._ElementTree)
 
         self.lxml = file
         self.doctype = self.lxml.docinfo.doctype
+
         self.dtd = dtd or self.lxml.docinfo.externalDTD
         self.source_url = self.lxml.docinfo.URL
         self.public_id = self.lxml.docinfo.public_id
         self.encoding = self.lxml.docinfo.encoding
-        self.sch_schemas = list(sch_schemas) if sch_schemas else []
 
-        # Load python based validation pipeline
-        self.ppl = checks.StyleCheckingPipeline()
+        if style_validators:
+            self.style_validators = list(style_validators)
+        else:
+            self.style_validators = []
 
     @classmethod
     def parse(cls, file, no_doctype=False, sps_version=None,
@@ -182,16 +250,22 @@ class XMLValidator(object):
         # can raise exception
         sps_version = sps_version or _init_sps_version(et, supported_sps_versions)
 
-        # get the right Schematron schema based on the value of ``sps_version``
+        # get the right Schematron validator based on the value of ``sps_version``
         # and then mix it with the list of schemas supplied by the user.
-        sch_schemas = [StdSchematron(sps_version)]
+        LOGGER.info('auto-loading style validations for version "%s"', sps_version)
+        style_validators = [
+                SchematronValidator.from_catalog(sps_version),
+                PyValidator(),  # the python based validation pipeline
+        ]
         if extra_sch_schemas:
-            sch_schemas += list(extra_sch_schemas)
+            style_validators += [SchematronValidator(sch) 
+                                 for sch in list(extra_sch_schemas)]
 
         allowed_public_ids = _get_public_ids(sps_version)
 
         # DOCTYPE declaration must be present by default. This behaviour can
         # be changed by the `no_doctype` arg.
+        LOGGER.info('fetching the DOCTYPE declaration')
         doctype = et.docinfo.doctype
         if not doctype and not no_doctype:
             raise exceptions.XMLDoctypeError(
@@ -199,11 +273,19 @@ class XMLValidator(object):
 
         # if there exists a DOCTYPE declaration, ensure its PUBLIC-ID is
         # supported.
+        LOGGER.info('fetching the PUBLIC-ID in DOCTYPE declaration')
         public_id = et.docinfo.public_id
         if doctype and public_id not in allowed_public_ids:
             raise exceptions.XMLDoctypeError('invalid DOCTYPE public id')
 
-        return cls(et, sch_schemas=sch_schemas, **kwargs)
+        return cls(et, style_validators=style_validators, **kwargs)
+
+    @property
+    def dtd_validator(self):
+        if self.dtd:
+            return DTDValidator(self.dtd)
+        else:
+            return None
 
     @utils.cachedmethod
     def validate(self):
@@ -211,31 +293,11 @@ class XMLValidator(object):
 
         Returns a tuple comprising the validation status and the errors list.
         """
-        if self.dtd is None:
+        if self.dtd_validator is None:
             raise exceptions.UndefinedDTDError('cannot validate (DTD is not set)')
 
-        def make_error_log():
-            return [style_errors.SchemaStyleError(err) 
-                    for err in self.dtd.error_log]
-
-        result = self.dtd.validate(self.lxml)
-        errors = make_error_log()
-
-        return result, errors
-
-    def _validate_sch(self, schema):
-        """Validate the source XML against SPS-Style Schematron.
-
-        Returns a tuple comprising the validation status and the errors list.
-        """
-        def make_error_log(schematron):
-            err_log = schematron.error_log
-            return [style_errors.SchematronStyleError(err) for err in err_log]
-
-        result = schema.validate(self.lxml)
-        errors = make_error_log(schema)
-
-        return result, errors
+        result_tuple = self.dtd_validator.validate(self.lxml)
+        return result_tuple
 
     @utils.cachedmethod
     def validate_style(self):
@@ -243,12 +305,9 @@ class XMLValidator(object):
 
         Returns a tuple comprising the validation status and the errors list.
         """
-        # run python based checks
-        errors = next(self.ppl.run(self.lxml, rewrap=True))
-
-        # now run all the schematron schemas
-        for schema in self.sch_schemas:
-            errors += self._validate_sch(schema)[1]
+        errors = []
+        for validator in self.style_validators:
+            errors += validator.validate(self.lxml)[1]
 
         result = not bool(errors)